move ButtonLink to single folder
